+++ saleor/plugins/manager.py
@@ -1457,11 +1457,12 @@ class PluginsManager(PaymentInterface):
     def list_payment_gateways(
         self,
         currency: Optional[str] = None,
-        checkout: Optional["Checkout"] = None,
+        checkout_info: Optional["CheckoutInfo"] = None,
+        checkout_lines: Optional[Iterable["CheckoutLineInfo"]] = None,
         channel_slug: Optional[str] = None,
         active_only: bool = True,
     ) -> List["PaymentGateway"]:
-        channel_slug = checkout.channel.slug if checkout else channel_slug
+        channel_slug = checkout_info.channel.slug if checkout_info else channel_slug
         plugins = self.get_plugins(channel_slug=channel_slug, active_only=active_only)
         payment_plugins = [
             plugin for plugin in plugins if "process_payment" in type(plugin).__dict__
@@ -1472,7 +1473,10 @@ class PluginsManager(PaymentInterface):
         for plugin in payment_plugins:
             gateways.extend(
                 plugin.get_payment_gateways(
-                    currency=currency, checkout=checkout, previous_value=None
+                    currency=currency,
+                    checkout_info=checkout_info,
+                    checkout_lines=checkout_lines,
+                    previous_value=None,
                 )
             )
         return gateways